==> notebooks/notebook1.ipynb <==
from pathlib import Path
import requests
import os

def download_one_file_of_raw_data(year: int, month: int):
    url = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"
    response = requests.get(url)

    if response.status_code == 200:
        directory = "../data/raw/"
        os.makedirs(directory, exist_ok=True)
        path = f"../data/raw/rides_{year}-{month:02d}.parquet"
        open(path, "wb").write(response.content)
        return path
    else:
        raise Exception(f"Failed to download {url}")

download_one_file_of_raw_data(2022, 1)

import pandas as pd

data = pd.read_parquet("../data/raw/rides_2022-01.parquet")

data.size

data.head()

rides = data[['tpep_pickup_datetime', 'PULocationID']]

rides.head()

rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime',
    'PULocationID': 'pickup_location_id'
}, inplace=True)

rides.head()

rides['pickup_datetime'].describe()

rides = rides[rides.pickup_datetime >= '2022-01-01']
rides = rides[rides.pickup_datetime < '2022-02-01']
rides['pickup_datetime'].describe()

os.makedirs('../data/transformed', exist_ok=True)
rides.to_parquet('../data/transformed/rides_2022_01.parquet')

from src.paths import TRANSFORMED_DATA_DIR
