==> notebooks/notebook2.ipynb <==
import pandas as pd

rides = pd.read_parquet("../data/transformed/rides_2022_01.parquet")

rides.head()

rides["pickup_hour"] = rides['pickup_datetime'].dt.floor('h')

rides.head()

app_rides = rides.groupby(["pickup_hour", "pickup_location_id"]).size().reset_index()
app_rides.rename(columns = {0: 'rides'}, inplace = True)

app_rides.head()

unique_days = app_rides["pickup_hour"].dt.floor('D').unique()

# Generate all hours of the day for each unique date
full_range = pd.concat([
    pd.DataFrame({"pickup_hour": pd.date_range(start=day, end=day + pd.Timedelta(days=1), freq="H")[:-1]})
    for day in unique_days
], ignore_index=True)

full_range

from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:

    location_ids = agg_rides["pickup_location_id"].unique()
    unique_days = agg_rides["pickup_hour"].dt.floor('D').unique()

# Generate all hours of the day for each unique date
    full_range = pd.concat([
        pd.DataFrame({"pickup_hour": pd.date_range(start=day, end=day + pd.Timedelta(days=1), freq="H")[:-1]})
        for day in unique_days
    ], ignore_index=True)
    
    output = pd.DataFrame()

    for location_id in tqdm(location_ids):
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour', 'rides']]

        agg_rides_i.set_index("pickup_hour", inplace=True)
        agg_rides_i.index = pd.to_datetime(agg_rides_i.index)

        agg_rides_i = agg_rides_i.reindex(full_range['pickup_hour'], method="ffill")

        agg_rides_i["pickup_location_id"] = location_id
        output = pd.concat([output, agg_rides_i])

    output = output.reset_index().rename(columns={"index": "pickup_hour"})
    output['rides'] = output['rides'].fillna(0)

    return output

agg_rides_all_slots = add_missing_slots(app_rides)

from typing import Optional, List
import plotly.express as px

def plot_rides(
        rides: pd.DataFrame,
        locations: Optional[List[int]] = None
):
    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y="rides",
        color="pickup_location_id"
    )

    # fig.show(
    #     rides_to_plot,
    #     x="pickup_hour",
    #     y="rides",
    #     color="pickup_location_id",
    #     template = 'none'
    # )

    fig.show()

plot_rides(agg_rides_all_slots,locations = [100])
