==> notebooks/notebook3.ipynb <==
import pandas as pd

ts_data = pd.read_parquet("../data/transformed/ts_data_2022_01.parquet")
ts_data.head()

ts_data_one_location = ts_data.loc[ts_data.pickup_location_id == 43, :].reset_index(drop=True)
ts_data_one_location.head(25)

def get_cutoff_indices(
        data: pd.DataFrame,
        n_features: int,
        step_size: int
) -> list:
    stop_position = len(data) -1
    subseq_first_idx = 0
    subseq_mid_index = n_features
    subseq_last_index = n_features + 1
    indices = []
    while subseq_last_index <= stop_position:
        indices.append((subseq_first_idx, subseq_mid_index, subseq_last_index))
        subseq_first_idx += step_size
        subseq_mid_index += step_size
        subseq_last_index += step_size

    return indices

n_features = 24
step_size = 1

indices = get_cutoff_indices(ts_data_one_location, n_features, step_size)

indices[:5]

import numpy as np

n_examples = len(indices)
x = np.ndarray(shape = (n_examples, n_features), dtype=np.float32)
y = np.ndarray(shape = (n_examples), dtype=np.float32)
pickup_hours = []
for i, idx in enumerate(indices):
    x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
    y[i] = ts_data_one_location.iloc[idx[1]:idx[2]]['rides'].values
    pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

x.shape, y.shape

features_one_location = pd.DataFrame(
    x,
    columns = [f'rides_previous_{i+1}_hour' for i in reversed(range(n_features))]
)
features_one_location

targets_one_location = pd.DataFrame(
    y,
    columns = [f'target_rides_next_hour']
)
targets_one_location

from tqdm import tqdm

def transform_ts_data_into_features_and_target(
        ts_data: pd.DataFrame,
        n_sez_len: int,
        step_size: int
) -> pd.DataFrame:
    assert set(ts_data.columns) == {'pickup_hour', 'rides', 'pickup_location_id'}

    location_ids = ts_data['pickup_location_id'].unique()
    features = pd.DataFrame()
    targets = pd.DataFrame()

    for location_id in tqdm(location_ids):
        ts_data_one_location = ts_data.loc[ts_data.pickup_location_id == location_id,
                                            ['pickup_hour', 'rides']
                                            ]
        indices = get_cutoff_indices(ts_data_one_location, n_sez_len, step_size)
        n_examples = len(indices)
        x = np.ndarray(shape = (n_examples, n_sez_len), dtype=np.float32)
        y = np.ndarray(shape = (n_examples), dtype=np.float32)
        pickup_hours = []
        for i, idx in enumerate(indices):
            x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
            y[i] = ts_data_one_location.iloc[idx[1]:idx[2]]['rides'].values
            pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

        features_one_location = pd.DataFrame(
            x,
            columns = [f'rides_previous_{i+1}_hour' for i in reversed(range(n_sez_len))]
        )

        features_one_location['pickup_hour'] = pickup_hours
        features_one_location['pickup_location_id'] = location_id

        targets_one_location = pd.DataFrame(
            y,
            columns = [f'target_rides_next_hour']
        )

        features = pd.concat([features, features_one_location])
        targets = pd.concat([targets, targets_one_location])
    
    features.reset_index(drop=True, inplace=True)
    targets.reset_index(drop=True, inplace=True)

    return features, targets['target_rides_next_hour']

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    n_sez_len = 24*7*1,
    step_size = 24
)
print(f'{features.shape = }')
print(f'{targets.shape = }')

features.head()

targets.head()
